+++ src/gui/images/bruj.c
