+++ src/gui/images/compass.c
