+++ src/gui/images/navigation.c
