+++ src/gui/images/speed.c
