+++ src/gui/images/zoom.c
