+++ src/gui/screens/Button_Bar/events/button_bar.h
@@ -1,0 +1,18 @@
+/**
+ * @file button_bar.h
+ * @author Jordi Gauchía (jgauchia@jgauchia.com)
+ * @brief  Button Bar events
+ * @version 0.1.6
+ * @date 2023-06-14
+ */
+
+
+/**
+ * @brief Settings Button event
+ * 
+ * @param event 
+ */
+void settings(lv_event_t *event)
+{
+    log_v("Settings");
+}
