+++ src/gui/screens/Splash/splash_scr.h
